NP-16765 Add doi accordion
