+++ src/pages/public_registration/PublicRegistrationStatusBar.tsx
@@ -1,19 +1,17 @@
 import { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { Box, Button, CircularProgress, DialogActions, TextField, Typography } from '@mui/material';
+import { Box, CircularProgress, Typography } from '@mui/material';
 import CloudUploadIcon from '@mui/icons-material/CloudUpload';
 import CloseIcon from '@mui/icons-material/Close';
 import CheckIcon from '@mui/icons-material/Check';
-import LocalOfferIcon from '@mui/icons-material/LocalOffer';
 import { useTranslation } from 'react-i18next';
 import { validateYupSchema, yupToFormErrors } from 'formik';
 import { LoadingButton } from '@mui/lab';
 import { RootState } from '../../redux/store';
 import { PublicRegistrationContentProps } from './PublicRegistrationContent';
-import { Modal } from '../../components/Modal';
 import { setNotification } from '../../redux/notificationSlice';
 import { RegistrationStatus, Registration } from '../../types/registration.types';
-import { addTicketMessage, createTicket, updateTicketStatus } from '../../api/registrationApi';
+import { createTicket, updateTicketStatus } from '../../api/registrationApi';
 import { registrationValidationSchema } from '../../utils/validation/registration/registrationValidation';
 import { getFirstErrorTab, getTabErrors, TabErrors } from '../../utils/formik-helpers';
 import { dataTestId } from '../../utils/dataTestIds';
@@ -26,7 +24,6 @@ import { BackgroundDiv } from '../../components/styled/Wrappers';
 enum LoadingState {
   None,
   Publish,
-  RequestDoi,
   RejectDoi,
   ApproveDoi,
   ApprovePublishRequest,
@@ -44,12 +41,8 @@ export const PublicRegistrationStatusBar = ({
   const dispatch = useDispatch();
   const { t } = useTranslation();
   const user = useSelector((store: RootState) => store.user);
-  const { doi } = registration;
 
-  const [messageToCurator, setMessageToCurator] = useState('');
-  const [openRequestDoiModal, setOpenRequestDoiModal] = useState(false);
   const [isLoading, setIsLoading] = useState(LoadingState.None);
-  const toggleRequestDoiModal = () => setOpenRequestDoiModal((state) => !state);
   const [tabErrors, setTabErrors] = useState<TabErrors>();
 
   const [registrationTicketCollection, isLoadingRegistrationTicketCollection] = useFetch<TicketCollection>({
@@ -65,32 +58,6 @@ export const PublicRegistrationStatusBar = ({
   const pendingPublishingRequestTicket = registrationTickets.find(
     (ticket) => ticket.type === 'PublishingRequest' && ticket.status === 'Pending'
   );
-
-  const sendDoiRequest = async () => {
-    setIsLoading(LoadingState.RequestDoi);
-    const message = isPublishedRegistration ? messageToCurator : t('registration.public_page.reserve_doi_message');
-
-    const createDoiRequestResponse = await createTicket(registration.id, 'DoiRequest');
-    if (isErrorStatus(createDoiRequestResponse.status)) {
-      dispatch(setNotification({ message: t('feedback.error.create_doi_request'), variant: 'error' }));
-      setIsLoading(LoadingState.None);
-    } else if (isSuccessStatus(createDoiRequestResponse.status)) {
-      const ticketId = createDoiRequestResponse.data.id;
-      // Add message
-      if (ticketId && message) {
-        await addTicketMessage(ticketId, message);
-        // No need to show potential error message, since Ticket with actual DoiRequest is created anyway
-      }
-      // TODO: Adding DOI can take some extra time, so wait 10 sec before refetching
-      setTimeout(() => {
-        if (openRequestDoiModal) {
-          toggleRequestDoiModal();
-        }
-        dispatch(setNotification({ message: t('feedback.success.doi_request_sent'), variant: 'success' }));
-        refetchRegistration();
-      }, 10_000);
-    }
-  };
 
   const updatePendingDoiRequest = async (status: TicketStatus) => {
     if (pendingDoiRequestTicket) {
@@ -181,7 +148,6 @@ export const PublicRegistrationStatusBar = ({
   const registrationIsValid = !tabErrors || firstErrorTab === -1;
 
   const isCurator = userIsRegistrationCurator(user, registration);
-  const hasNvaDoi = !!doi;
   const isPublishedRegistration = registration.status === RegistrationStatus.Published;
 
   return (
@@ -219,52 +185,6 @@ export const PublicRegistrationStatusBar = ({
                 {t('common.publish')}
               </LoadingButton>
             )}
-
-          {!hasNvaDoi && (
-            <>
-              <LoadingButton
-                variant="outlined"
-                endIcon={<LocalOfferIcon />}
-                loadingPosition="end"
-                loading={isLoading === LoadingState.RequestDoi}
-                data-testid={
-                  isPublishedRegistration
-                    ? dataTestId.registrationLandingPage.requestDoiButton
-                    : dataTestId.registrationLandingPage.reserveDoiButton
-                }
-                onClick={() => (isPublishedRegistration ? toggleRequestDoiModal() : sendDoiRequest())}>
-                {isPublishedRegistration
-                  ? t('registration.public_page.request_doi')
-                  : t('registration.public_page.reserve_doi')}
-              </LoadingButton>
-              <Modal
-                open={openRequestDoiModal}
-                onClose={toggleRequestDoiModal}
-                headingText={t('registration.public_page.request_doi')}
-                dataTestId={dataTestId.registrationLandingPage.requestDoiModal}>
-                <Typography paragraph>{t('registration.public_page.request_doi_description')}</Typography>
-                <TextField
-                  variant="outlined"
-                  multiline
-                  rows="4"
-                  fullWidth
-                  data-testid={dataTestId.registrationLandingPage.doiMessageField}
-                  label={t('registration.public_page.message_to_curator')}
-                  onChange={(event) => setMessageToCurator(event.target.value)}
-                />
-                <DialogActions>
-                  <Button onClick={toggleRequestDoiModal}>{t('common.cancel')}</Button>
-                  <LoadingButton
-                    variant="contained"
-                    data-testid={dataTestId.registrationLandingPage.sendDoiButton}
-                    onClick={sendDoiRequest}
-                    loading={isLoading === LoadingState.RequestDoi}>
-                    {t('common.send')}
-                  </LoadingButton>
-                </DialogActions>
-              </Modal>
-            </>
-          )}
 
           {isCurator ? (
             isLoadingRegistrationTicketCollection ? (
